requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
-e git+ssh://git@github.com/lociere/Cradle_Selrena.git@14390ce09d50dcedcb8e07ce71e1ff943c6a8c41#egg=cradle_selrena
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.9
langchain-core==1.2.9
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4
langsmith==0.6.9
loguru==0.7.3
MouseInfo==0.1.3
numpy==2.4.2
openai==2.17.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pillow==12.1.0
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
PySide6==6.10.2
PySide6_Addons==6.10.2
PySide6_Essentials==6.10.2
python-dotenv==1.2.1
pytweening==1.2.0
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
shiboken6==6.10.2
sniffio==1.3.1
sounddevice==0.5.5
llama-cpp-python>=0.3.0
huggingface_hub>=0.20.0
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
webrtcvad-wheels==2.0.14
win32_setctime==1.2.0
xxhash==3.6.0
zstandard==0.25.0
